LayoutTests/inspector/debugger-step-in.html

<html>
<head>
<script src="../http/tests/inspector/inspector-test.js"></script>
<script src="../http/tests/inspector/debugger-test.js"></script>
<script>

function d() {
    var i = 10;
}

function testFunction() {
    debugger;
    d();
    notifyDone();
}

function doit()
{
    evaluateInWebInspector("frontend_testStepIn", dumpArray);
}

function frontend_testStepIn(testController)
{
    testController.waitUntilDone();

    WebInspector.showPanel("scripts");
    frontend_ensureDebuggerEnabled(debuggerWasEnabledCallback);

    function debuggerWasEnabledCallback()
    {
        testController.results.push("Debugger was enabled.");
        frontend_evaluateInConsole("setTimeout(testFunction, 0)", function(result)
        {
            testController.results.push("Set timer for test function.");
        });
    }

    var stepCount = 0;
    frontend_addSniffer(WebInspector, "pausedScript", function(callFrames)
    {
        testController.results.push("Script execution paused.");
        frontend_captureStackTrace(callFrames, testController);
        if (stepCount == 2)
            frontend_completeDebuggerTest(testController);
        else {
            testController.results.push("Stepping into...");
            document.getElementById("scripts-step-into").click();
        }
        stepCount++;
    }, true);
}

</script>
</head>

<body onload="onload()">
<p>
Tests "step in" functionality in debugger.
</p>

</body>
</html>
